* master: (250 commits)
  src/sndfile.c: Check return of write_header()
  src/flac.c: Improve error handling/reporting
  Add AM_PROG_AR to configure.ac
  'rm -f' can be used on non-existent files.
  Make 'test-sndfile-metadata-set.py' py2/py3 compatible
  Test whether -D_FORTIFY_SOURCE needs to be added at all
  Refactor generation of test-suite
  Make programs optional
  Replace AC_HELP_STRING by AS_HELP_STRING
  Compile 'sndfile-regtest' only when sqlite3 is available
  Compile tests only when running 'make check'
  Remove 'INSTALL' file, as it is part of the cruft automake installs always
  tests/write_read_test.tpl: One more check
  src/sndfile.c: Improve validation in sf_read/write_*
  src/flac.c: Cleanup and validation improvements
  src/windows.c: Fix sf_wchar_open
  src/test_binheader_writef.c: Initialize sf_private
  tests/chunk_test.c: Add test for large free chunks
  Scripts/static-deps-build.mk: Restore file
  src/ : Follow changes to header buffer
  ...

Conflicts:
	src/aiff.c
	src/common.c
	src/common.h
	src/wav.c

+++ src/ALAC/ag_dec.c
@@ -26,14 +26,14 @@
 	Copyright:	(c) 2001-2011 Apple, Inc.
 */
 
-#include "aglib.h"
-#include "ALACBitUtilities.h"
-#include "ALACAudioTypes.h"
-
 #include <math.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+
+#include "aglib.h"
+#include "ALACBitUtilities.h"
+#include "ALACAudioTypes.h"
 
 #define CODE_TO_LONG_MAXBITS	32
 #define N_MAX_MEAN_CLAMP		0xffff
@@ -306,7 +306,7 @@ int32_t dyn_decomp (AGParamRecPtr params, BitBuffer * bitstream, int32_t * pc, i
 		// least significant bit is sign bit
 		{
 			uint32_t	ndecode = n + zmode ;
-			int32_t		multiplier = - (ndecode & 1) ;
+			int32_t		multiplier = - (int) (ndecode & 1) ;
 
 			multiplier |= 1 ;
 			del = ((ndecode+1) >> 1) * (multiplier) ;
